:boom: Rename or remove internal modules
Some modules should not be exposed so renamed them with prefix `_` or
removed.
