+++ bufname/_utils_test.ts
@@ -1,5 +1,5 @@
 import { assertEquals } from "https://deno.land/std@0.211.0/assert/mod.ts";
-import { decode, encode } from "./utils.ts";
+import { decode, encode } from "./_utils.ts";
 
 Deno.test("encode does nothing on alphabet characters", () => {
   const src = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ_";
