+++ mapping/_parser_test.ts
@@ -3,7 +3,7 @@ import {
   assertThrows,
 } from "https://deno.land/std@0.211.0/assert/mod.ts";
 import { Mapping } from "./types.ts";
-import { parse } from "./parser.ts";
+import { parse } from "./_parser.ts";
 
 const testcases: [string, Mapping][] = [
   ["n  <Space>hp    @<Plug>(test-denops-std)", {
